beardog_workflows::workflows::notification

Module adapters

Notification adapter plugin system

This module provides a framework for users to create custom notification adapters for any service they use - Matrix, Signal, custom internal systems, etc.

Structs

AdapterConfigTemplate

Configuration template for an adapter type

AdapterRegistry

Plugin-based adapter registry

ConfigField

Configuration field definition

DiscordFactory

Discord adapter factory

EmailFactory

Email adapter factory

GenericWebhookFactory

Generic webhook adapter factory for any webhook service

MatrixFactory

Matrix adapter factory (for Matrix/Element users)

SignalFactory

Signal adapter factory (for Signal messenger users)

SlackFactory

Slack adapter factory

SmsFactory

SMS adapter factory

TeamsFactory

Microsoft Teams adapter factory

Enums

ConfigFieldType

Configuration field types

Traits

AdapterFactory

Factory trait for creating adapters from configuration